Fix python vault
